backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
pymysql==1.1.0
reportlab==4.0.4
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
bcrypt==4.0.1
flask-mail==0.9.1